add test data in csv and script for generating/model+data
