+++ test/unit/data/prediction_intervals/test_ts.csv
@@ -1,0 +1,20 @@
+-1.128422993953936526e+01
+-1.152021043390404742e+01
+-1.211148841442386903e+01
+-1.132843192119836040e+01
+-1.182764295524443909e+01
+-1.302490946891641954e+01
+-1.380193732063798251e+01
+-1.289498486022025681e+01
+-1.170175382248446283e+01
+-1.297384314724285659e+01
+-1.383549595641619412e+01
+-1.329118421073531309e+01
+-1.372101702763957753e+01
+-1.168317495413726625e+01
+-1.236394109552265341e+01
+-1.411815241978704982e+01
+-1.310470807313166475e+01
+-1.132397062058253923e+01
+-1.336359043438720562e+01
+-1.080284556571273313e+01
