+++ test/unit/data/prediction_intervals/train_ts.csv
@@ -1,0 +1,200 @@
+1.070815027042121947e+00
+-1.826382209949657165e+00
+-8.791555968617592587e-01
+-5.166477364718884857e+00
+-2.952528179743950432e+00
+-2.450943651537444090e+00
+-3.297866873616331684e+00
+-1.694114995000752577e+00
+-3.122263962371539581e+00
+-1.594010387496368475e+00
+-4.658833012071235125e-01
+-4.350166773238214546e+00
+-4.225530458248380583e+00
+-4.012067136888918029e+00
+-1.715175598570751303e+00
+-2.419860164259397983e+00
+-3.116685276015697337e+00
+-2.136080415833144563e+00
+-1.172360424047367289e+00
+-4.764722385955488804e+00
+-2.745556358030226907e+00
+-1.871453494273818130e-01
+-2.656308984686469810e+00
+-1.079807226885942217e+00
+-3.304026438811739297e+00
+-3.892979303587019047e+00
+-1.983976911507412400e+00
+-1.588383990386246003e+00
+-4.447771906599648162e+00
+-1.957011419215115566e+00
+-4.760434694935002042e+00
+-3.613972840916588769e+00
+-1.446483273551444126e+00
+-2.726623461750607547e+00
+-2.253307726469993888e+00
+-4.310491423375649944e+00
+-6.096411474990624413e+00
+-2.897605989817330308e+00
+-2.020557277721808376e+00
+-4.914039698304103254e+00
+-3.054521504114211083e+00
+-6.142713265304799108e+00
+-5.519369579681758786e+00
+-4.586733317745347982e+00
+-4.509350004538251078e+00
+-5.002563260031285708e+00
+-5.196169001414192579e+00
+-4.456275042740212022e+00
+-7.866217106305762208e+00
+-3.829000894750885120e+00
+-6.228237151368658076e+00
+-7.368876729353520005e+00
+-4.812364571410225444e+00
+-5.754475981088992498e+00
+-7.531430211799492014e+00
+-7.596118042930667613e+00
+-7.688206289022272877e+00
+-7.865829182828834476e+00
+-8.477598164033254946e+00
+-8.676673536565555622e+00
+-7.621762758524747383e+00
+-9.520732099169331164e+00
+-7.924275723732092658e+00
+-1.072884570814186311e+01
+-1.028123997858416239e+01
+-1.052489632465103675e+01
+-1.056535643201229391e+01
+-1.002815565968210088e+01
+-1.136647040283390098e+01
+-8.889855573771097141e+00
+-9.023407153498904876e+00
+-1.117689659188791929e+01
+-1.197435980851679993e+01
+-1.338394698200238331e+01
+-1.478456429580954534e+01
+-1.196921948676248526e+01
+-1.365669808480233627e+01
+-1.292279266561338424e+01
+-1.355845990954069080e+01
+-1.184905960135443515e+01
+-1.457462414630789205e+01
+-1.371037887281861778e+01
+-1.560478860102606724e+01
+-1.325885209879948690e+01
+-1.447816779334771020e+01
+-1.456824099271986128e+01
+-1.454929776759069071e+01
+-1.501462004892678870e+01
+-1.303351328663105413e+01
+-1.474985649312778335e+01
+-1.617095448508390021e+01
+-1.404843356715485925e+01
+-1.404417793947498261e+01
+-1.584908871112568463e+01
+-1.486946734160240524e+01
+-1.453102514983551430e+01
+-1.609075497548366229e+01
+-1.503415980565021925e+01
+-1.504779950896583607e+01
+-1.553771360073981178e+01
+-1.506003896606762815e+01
+-1.417922407337037782e+01
+-1.447973689775040285e+01
+-1.394493715578155246e+01
+-1.494183757467856744e+01
+-1.385482093527867065e+01
+-1.346522723438532410e+01
+-1.440107959537469107e+01
+-1.302151534247322928e+01
+-1.546189764140142486e+01
+-1.288167907237909127e+01
+-1.340553296084867974e+01
+-1.286155442794832915e+01
+-1.321901504062090105e+01
+-1.246997833718980075e+01
+-1.345077026281832921e+01
+-1.125333470504893896e+01
+-1.219095217973776002e+01
+-1.180183547269165878e+01
+-1.147798227946104355e+01
+-1.019742487542985465e+01
+-9.997355369112511880e+00
+-9.737289593196859272e+00
+-1.105004314690799916e+01
+-9.189826420852273969e+00
+-9.652196796672862789e+00
+-1.114055026611043964e+01
+-9.513749228534175373e+00
+-8.671279334443907416e+00
+-8.615009067504008300e+00
+-9.128699528691310405e+00
+-9.843800831023628461e+00
+-8.338695792214089408e+00
+-7.156853194762057946e+00
+-7.771601853400734505e+00
+-7.648880433052449135e+00
+-5.910080997934017866e+00
+-6.432628816066976896e+00
+-5.164910247660010612e+00
+-6.797407418342773511e+00
+-5.035711742370839517e+00
+-7.645482929209827638e+00
+-7.613997684757441320e+00
+-6.683591954441935989e+00
+-4.388921525832794934e+00
+-6.147375215269021531e+00
+-6.438021841034564652e+00
+-5.046830987348914555e+00
+-4.705050682138429252e+00
+-6.533573609780650493e+00
+-4.732445790996490764e+00
+-4.775749717912104231e+00
+-6.538607174418573642e+00
+-4.458216964385662173e+00
+-4.342792635250403599e+00
+-4.843559746933492782e+00
+-4.022318455390703207e+00
+-2.585291615237237828e+00
+-3.244073203638021319e+00
+-4.227808805319490659e+00
+-3.686885483936922725e+00
+-6.328235319770905143e+00
+-3.147145015480884211e+00
+-6.974551171626008461e+00
+-5.705739895196317413e+00
+-4.939255149073301432e+00
+-4.640035483494992619e+00
+-4.081702076061460183e+00
+-5.135450023030758260e+00
+-5.087263465415308161e+00
+-3.416583940411910714e+00
+-4.020037046363508892e+00
+-5.751047066231446792e+00
+-6.685674169064822259e+00
+-4.160621183009364010e+00
+-5.239118250562528445e+00
+-6.204430351366288399e+00
+-6.557241757745540411e+00
+-6.705041233121142419e+00
+-5.654421981916049944e+00
+-5.662549624305676765e+00
+-9.241093412504438831e+00
+-7.519097534962567941e+00
+-6.348163736577520311e+00
+-7.146099702105225049e+00
+-9.515717842289889816e+00
+-9.302286184323616425e+00
+-7.814139457712707681e+00
+-9.211438536045083225e+00
+-8.302837448788073615e+00
+-7.083324284426011808e+00
+-9.321808850146599212e+00
+-1.096368395498414117e+01
+-1.017320724421113631e+01
+-9.235070713084757799e+00
+-1.062701286898984243e+01
+-9.955022106980344176e+00
+-9.905564327127716950e+00
+-1.185222438709485537e+01
+-1.166515347864083374e+01
